Moving auth logic into a base
